Library Reference

Syntax

GATlab.Syntax.Scopes.Binding — Type

Binding{T, Sig}

A binding associates some T-typed value to a name and possibly some aliases, disambiguated by a signature in Sig in the case of overloading.

primary is an optional distinguished element of aliases value is the element sig is a way of uniquely distinguishing this element from others with the same name (for example, ⊗ : Ob x Ob -> Ob and Hom x Hom -> Hom)

GATlab.Syntax.Scopes.Context — Type

A Context is anything which contains an ordered list of scopes.

Scopes within a context are referred to by level, which is their index within this list.

Contexts should overload:

getscope(c::Context, level::Int) -> Scope

nscopes(c::Context) -> Int

hastag(c::Context, tag::ScopeTag) -> Bool

hasname(c::Context, name::Symbol) -> Bool

getlevel(c::Context, tag::ScopeTag) -> Int

getlevel(c::Context, name::Symbol) -> Int

GATlab.Syntax.Scopes.HasScope — Type

An abstract type for wrappers around a single scope.

Must overload

getscope(hs::HasScope) -> Scope

GATlab.Syntax.Scopes.HasScopeList — Type

A type for things which contain a scope list.

Notably, GATs contain a scope list.

Must implement:

getscopelist(hsl::HasScopeList) -> ScopeList

GATlab.Syntax.Scopes.Ident — Type

Ident

An identifier.

tag refers to the scope that this Ident is bound in lid indexes the scope that Ident is bound in name is an optional field for the sake of printing. A variable in a scope might be associated with several names

GATlab.Syntax.Scopes.LID — Type

A LID (Local ID) indexes a given scope.

Currently, scopes assign LIDs sequentially – this is not a stable guarantee however, and in theory scopes could have "sparse" LIDs.

GATlab.Syntax.Scopes.Reference — Type

Reference

A path of identifiers. We optimize for the (frequent) case where there is only one identifier by making this a linked list.

GATlab.Syntax.Scopes.Scope — Type

Scope{T, Sig}

In GATlab, we handle overloading and shadowing with a notion of scope. Anything which binds variables introduces a scope, for instance a @theory declaration or a context. For example, a scope with 3 elements:

x::Int = 3
y::String = "hello"
x::String = "ex"

This is a valid scope even though there are name collisions, because the signature (in this case, a datatype) disambiguates. Of course, it may not be wise to disambiguate by type, because it is not always possible to infer expected type. In general, one should pick something that can be inferred, and Nothing is always a reasonable choice, which disallows any name collisions.

GATlab.Syntax.Scopes.ScopeTag — Type

The tag that makes reference to a specific scope possible.

GATlab.Syntax.Scopes.ScopeTagError — Type

ScopeTagError

An error to throw when an identifier has an unexpected scope tag

Base.:+ — Method

Deterministically combine two scope tags into a third

GATlab.Syntax.Scopes.equiv — Method

Compare two scopes, ignoring the difference in the top-level scope tag.

GATlab.Syntax.Scopes.getlid — Method

Determine the level of a binding given the name and possibly the signature

GATlab.Syntax.Scopes.hasident — Method

hasident checks whether an identifier with specified data exists, by attempting to create it and returning whether or not that attempt failed.

GATlab.Syntax.Scopes.ident — Method

ident creates an Ident from a context and some partial data supplied as keywords.

Keywords arguments:

tag::Union{ScopeTag, Nothing}. The tag of the scope that the Ident is in.

name::Union{Symbol, Nothing}. The name of the identifier.

lid::Union{LID, Nothing}. The lid of the identifier within its scope.

level::Union{Int, Nothing}. The level of the scope within the context.

sig::Any. The signature of the identifier, to disambiguate between multiple identifiers with the same name within the same scope.

strict::Bool. If strict is true, throw an error if not found, else return nothing.

isunique::Bool. If isunique is true, then don't use the signature to disambiguate, instead fail if their are multiple identifiers with the same name in a scope.

GATlab.Syntax.Scopes.rename — Method

rename(tag::ScopeTag, replacements::Dict{Symbol, Symbol}, x::T) where {T} -> T

Recurse through the structure of x, and change any name n in scope tag to get(replacements, n, n). Overload this function on structs that have names in them.

GATlab.Syntax.Scopes.retag — Method

retag(replacements::Dict{ScopeTag, ScopeTag}, x::T) where {T} -> T

Recurse through the structure of x, swapping any instance of a ScopeTag t with get(replacements, t, t). Overload this function on structs that have ScopeTags within them.

GATlab.Syntax.Scopes.unsafe_addalias! — Method

unsafe_addalias! should only be used when unsafe_pushbinding! won't be called again, because new bindings won't get the old aliases.

GATlab.Syntax.GATs.Judgment — Type

Judgment

A judgment is either a type constructor, term constructor, or axiom; a GAT is composed of a list of judgments.

GATlab.Syntax.GATs.AbstractConstant — Type

We need this to resolve a mutual reference loop; the only subtype is Constant

GATlab.Syntax.GATs.AlgAxiom — Type

AlgAxiom

A declaration of an axiom

GATlab.Syntax.GATs.AlgSort — Type

AlgSort

A sort, which is essentially a type constructor without arguments ref must be reference to a AlgTypeConstructor

GATlab.Syntax.GATs.AlgSorts — Type

AlgSorts

A description of the argument sorts for a term constructor, used to disambiguate multiple term constructors of the same name.

GATlab.Syntax.GATs.AlgTerm — Type

AlgTerm

One syntax tree to rule all the terms. Head can be a reference to an AlgTermConstructor, to a Binding{AlgType, Nothing}, or simply an AbstractConstant

GATlab.Syntax.GATs.AlgTermConstructor — Type

AlgTermConstructor

A declaration of a term constructor

GATlab.Syntax.GATs.AlgType — Type

AlgType

One syntax tree to rule all the types. head must be reference to a AlgTypeConstructor

GATlab.Syntax.GATs.AlgTypeConstructor — Type

AlgTypeConstructor

A declaration of a type constructor

GATlab.Syntax.GATs.Constant — Type

Constant

A Julia value in an algebraic context. Checked elsewhere.

GATlab.Syntax.GATs.GAT — Type

GAT

A generalized algebraic theory. Essentially, just consists of a name and a list of GATSegments, but there is also some caching to make access faster. Specifically, there is a dictionary to map ScopeTag to position in the list of segments, and there are lists of all of the identifiers for term constructors, type constructors, and axioms so that they can be iterated through faster.

GATs allow overloading but not shadowing.

GATlab.Syntax.GATs.GATSegment — Type

GATSegment

A piece of a GAT, consisting of a scope that binds judgments to names, possibly disambiguated by argument sorts.

GATlab.Syntax.GATs.JudgmentBinding — Type

JudgmentBinding

A binding of a judgment to a name and possibly a signature.

GATlab.Syntax.GATs.SortScope — Type

SortScope

A scope where variables are assigned to AlgSortss, and no overloading is permitted.

GATlab.Syntax.GATs.TermInCtx — Type

A term with an accompanying type scope, e.g.

(a,b)::R

a*(a+b)

GATlab.Syntax.GATs.TermInCtx — Method

Get the canonical term associated with a term constructor

GATlab.Syntax.GATs.TypeScope — Type

TypeScope

A scope where variables are assigned to AlgTypes, and no overloading is permitted.

GATlab.Syntax.ExprInterop.fromexpr — Method

Some expr may have already been bound (e.g. by an explicit context) and thus oughtn't be interpreted as default type.

GATlab.Syntax.ExprInterop.fromexpr — Method

Some expr may have already been bound (e.g. by an explicit context) and thus oughtn't be interpreted as default type.

GATlab.Syntax.ExprInterop.fromexpr — Method

Return nothing if the binding we parse has already been bound.

GATlab.Syntax.GATs.argcontext — Method

Local context of an AlgTermConstructor, including the arguments themselves

GATlab.Syntax.GATs.bind_localctx — Method

Take a term constructor and determine terms of its local context.

This function is mutually recursive with infer_type.

GATlab.Syntax.GATs.equations — Method

Get equations for a term or type constructor

GATlab.Syntax.GATs.equations — Method

Implicit equations defined by a context.

This function allows a generalized algebraic theory (GAT) to be expressed as an essentially algebraic theory, i.e., as partial functions whose domains are defined by equations.

References:

(Cartmell, 1986, Sec 6: "Essentially algebraic theories and categories with finite limits")

(Freyd, 1972, "Aspects of topoi")

This function gives expressions for computing each of the elements of context from the args, as well as checking that the args are well-typed.

Example:

equations({f::Hom(a,b), g::Hom(b,c)}, {a::Ob, b::Ob, c::Ob}, ThCategory)

waysofcomputing = Dict(a => [dom(f)], b => [codom(f), dom(g)], c => [codom(g)], f => [f], g => [g])

Algorithm:

Start from the arguments. We know how to compute each of the arguments; they are given. Each argument tells us how to compute other arguments, and also elements of the context

GATlab.Syntax.GATs.infer_type — Method

Infer the type of the term of a term. If it is not in context, recurse on its arguments. The term constructor's output type yields the resulting type once its localcontext variables are substituted with the relevant AlgTerms.

          (x,y,z)::Ob, p::Hom(x,y), q::Hom(y,z)

E.g. given ––––––––––––––––––– id(x)⋅(p⋅q)

             (a,b,c)::Ob, f::Hom(a,b), g::Hom(b,c)

and output type: –––––––––––––––––– Hom(a,c)

We first recursively find {id(x) => Hom(x,x), p⋅q => Hom(x,z)}. We ultimately want an AlgTerm for everything in the output type's context such that we can substitute into Hom(a,c) to get the final answer. It will help to also compute the AlgType for everything in the context. We work backwards, since we start by knowing {f => id(x)::Hom(x,x), g=> p⋅q :: Hom(x,z)}. For a b and c, we use equations which tell us, e.g., that a = dom(f). So we can grab the first argument of the type of f (i.e. grab x from Hom(x,x)).

GATlab.Syntax.GATs.normalize_decl — Method

axiom=true adds a ::default to exprs like f(a,b) ⊣ [a::A, b::B]

GATlab.Syntax.GATs.parsetypescope — Method

Keep track of variables already bound (e.g. in local context) so that they need not be redefined, e.g. compose(f,g::Hom(b,c)) ⊣ [(a,b,c)::Ob, f::Hom(a,b)] (If f were not defined in the local context, it would be parsed as default.)

GATlab.Syntax.GATs.sortcheck — Method

sortcheck(ctx::Context, t::AlgTerm)

Throw an error if a the head of an AlgTerm (which refers to a term constructor) has arguments of the wrong sort. Returns the sort of the term.

GATlab.Syntax.GATs.sortcheck — Method

sortcheck(ctx::Context, t::AlgType)

Throw an error if a the head of an AlgType (which refers to a type constructor) has arguments of the wrong sort.

GATlab.Syntax.GATs.substitute_term — Method

Replace idents with AlgTerms.

GATlab.Syntax.Presentations.Presentation — Type

A presentation has a set of generators, given by a TypeScope, and a set of equations among terms which can refer to those generators. Each element of eqs is a list of terms which are asserted to be equal.

GATlab.Syntax.ExprInterop.fromexpr — Method

Parse, e.g.:

(a,b,c)::Ob
f::Hom(a, b)
g::Hom(b, c)
h::Hom(a, c)
h′::Hom(a, c)
compose(f, g) == h == h′

GATlab.Syntax.ExprInterop.toexpr — Method

Context of presentation is the underlying GAT

GATlab.Syntax.ExprInterop.fromexpr — Function

fromexpr(c::Context, e::Any, T::Type) -> Union{T,Nothing}

Converts a Julia Expr into type T, in a certain scope.

GATlab.Syntax.ExprInterop.toexpr — Function

toexpr(c::Context, t) -> Any

Converts GATlab syntax into an Expr that can be read in via fromexpr to get the same thing. Crucially, the output of this will depend on the order of the scopes in c, and if read back in with a different c may end up with different results.

GATlab.Syntax.TheoryInterface.GAT_MODULE_LOOKUP — Constant

When we declare a new theory, we add the scope tag of its new segment to this dictionary pointing to the module corresponding to the new theory.

Models

GATlab.Models.ModelInterface.ImplementationNotes — Type

ImplementationNotes

Information about how a model implements a GATSegment. Right now, just the docstring attached to the @instance macro, but could contain more info in the future.

GATlab.Models.ModelInterface.implements — Method

implements(m::Model, tag::ScopeTag) -> Union{ImplementationNotes, Nothing}

If m implements the GATSegment referred to by tag, then return the corresponding implementation notes.

GATlab.Models.ModelInterface.parse_instance_body — Method

Parses the raw julia expression into JuliaFunctions

GATlab.Models.ModelInterface.qualify_function — Method

Add model kwarg (it shouldn't have it already) Qualify method name to be in theory module Add context kwargs if not already present

TODO: throw error if there's junk kwargs present already?

GATlab.Models.ModelInterface.typecheck_instance — Method

Throw error if missing a term constructor. Provides default instances for type constructors and type arguments, which return true or error, respectively.

GATlab.Models.ModelInterface.@instance — Macro

Usage:

struct TypedFinSetC <: Model{Tuple{Vector{Int}, Vector{Int}}}
  ntypes::Int
end

@instance ThCategory{Vector{Int}, Vector{Int}} (;model::TypedFinSetC) begin
  Ob(v::Vector{Int}) = all(1 <= j <= model.ntypes for j in v)
  Hom(f::Vector{Int}, v::Vector{Int}, w::Vector{Int}) =
     length(f) == length(v) && all(1 <= y <= length(w) for y in f)

  id(v::Vector{Int}) = collect(eachindex(v))
  compose(f::Vector{Int}, g::Vector{Int}) = g[f]

  dom(f::Vector{Int}; context) = context.dom
  codom(f::Vector{Int}; context) = context.codom
end

struct SliceCat{Ob, Hom, C <: Model{Tuple{Ob, Hom}}} <: Model{Tuple{Tuple{Ob, Hom}, Hom}}
  c::C
end

@instance ThCategory{Tuple{Ob, Hom}, Hom} (;model::SliceCat{Ob, Hom, C}) where {Ob, Hom, C<:Model{Tuple{Ob, Hom}}} begin
end

Utilities

GATlab.Util.MetaUtils — Module

General-purpose tools for metaprogramming in Julia.

GATlab.Util.MetaUtils.generate_docstring — Method

Wrap Julia expression with docstring.

GATlab.Util.MetaUtils.generate_function — Method

Generate Julia expression for function definition.

GATlab.Util.MetaUtils.parse_docstring — Method

Parse Julia expression that is (possibly) annotated with docstring.

GATlab.Util.MetaUtils.parse_function — Method

Parse Julia function definition into standardized form.

GATlab.Util.MetaUtils.parse_function_sig — Method

Parse signature of Julia function.

GATlab.Util.MetaUtils.replace_symbols — Method

Replace symbols occuring anywhere in a Julia expression.

GATlab.Util.MetaUtils.replace_symbols — Method

Replace symbols occurring anywhere in a Julia function (except the name).

GATlab.Util.MetaUtils.strip_lines — Method

Remove all LineNumberNodes from a Julia expression.