Library Reference

Gatlab.Util.Names.Name — Type

Names are used to label parts of a GAT.

They are used for both human input and output of a GAT, but are not used internally.

Gatlab.Util.Names.SymLit — Type

We have a symbol wrapper because we get symbols from parsing, and it is faster to compare symbols than it is to compare strings.

Gatlab.Syntax.Theories.AbstractTheory — Type

A type-level signifier for a particular theory, used to control dispatch and to pass around theory objects (which can't be type parameters) at the type level.

Structs which subtype AbstractTheory should always be singletons, and have theory defined on them.

Gatlab.Syntax.Theories.AnonTrmTag — Type

This can be used when there isn't a specific struct like Category.compose. Specific structs are preferred because they make reading backtraces easier.

Gatlab.Syntax.Theories.AnonTypTag — Type

This can be used when there isn't a specific struct like Category.Ob. Specific structs are preferred because they make reading backtraces easier.

Gatlab.Syntax.Theories.FullContext — Type

The full context for a Trm or Typ consists of both the list of judgments in the theory, and also the list of judgments in the context.

TODO: maybe we should have different terms for "context judgment" and "theory judgment"

Gatlab.Syntax.Theories.Lvl — Type

The tag of a Lvl takes up two bits and can be three things:

00: part of the theory 01: part of the context 10: part of the argument to a dependent context

Gatlab.Syntax.Theories.TrmTag — Type

This is used as a supertype for the tag types in a theory that correspond to term constructors.

Example:

module Category
struct compose <: TrmTag{3} end
end

Gatlab.Syntax.Theories.Typ — Type

The head of a type can never come from a context, only a theory, because it should point at a type constructor judgment.

Gatlab.Syntax.Theories.TypTag — Type

This is used as a supertype for the tag types in a theory that correspond to type constructors.

Example:

module Category
struct Ob <: TypTag{1} end
end

Gatlab.Syntax.Theories.gettheory — Function

Meant to be overloaded as

gettheory(::T) = ...

where T is a singleton struct subtyping AbstractTheory

Returns the @ref(Theory) associated to T.

Gatlab.Syntax.Theories.gettheory — Method

A convenience overload of theory

Gatlab.Syntax.Theories.lookup — Method

Get the Lvl corresponding to a Name. This is the most recent judgment with that name.

Gatlab.Syntax.TheoryMaps.TheoryMap — Method

Map a context in the domain theory into a context of the codomain theory

Gatlab.Syntax.TheoryMaps.TheoryMap — Method

Suppose dom(f) is [X,Y,Z,P,Q] and codom(f) is [XX,ϕ,ψ] suppose we have a term: P(a,Q(b,c)) ⊢ a::X,b::Y,c::Z i.e. 4({1},5({2},{3}))

f maps all sorts to XX f maps {P(x,y) ⊢ x::X,y::Y} to ϕ(ψ(y),x) i.e. 2(3({2}),{1}) and {Q(u,w) ⊢ u::Y,w::Z} to ψ(w) i.e. 3({2})

We should our term translate first to ϕ(ψ(y),x) i.e. 2(3({2}),{1})

and then substitute x (i.e. {1}) for the mapped first argument y (i.e. 5) for f(q(b,c)) i.e. ϕ(ψ(ψ(c)),x) 2(3(3({3})),{1})

So f(4({1},5({2},{3}))) = 2(3(3({3})),{1})

Gatlab.Syntax.Pushouts.pushout — Method

Pushout two inclusions. f A ↪ B g ↓ ↓ C->⌜D

D is a copy of B with the novel judgments of C added aftewards.

Gatlab.Syntax.Pushouts.substitute_level — Method

v is a FinFunction from old de bruijn levels to new ones. Its domain is the size of the theory of the judgment.

Gatlab.Syntax.Visualization.Sequent — Type

Intermediate representation of a judgment for pretty printing

Gatlab.Syntax.Visualization.ctx_dict — Method

Get the strings required to print the a context t1 that extends t2. E.g. [["a","b","c"]=>"Ob", ["f","g"]=>["Hom(a,b)"], ...]

Gatlab.Syntax.Visualization.show_inctx — Method

Show a debruijn level term in a theory + context i marks where the theory effectively ends (indices higher than this refer to the context).

Gatlab.Syntax.Visualization.show_inctx — Method

Show a debruijn level type in a theory + context

Gatlab.Logic.EGraphs.context — Method

This computes the inferred context for an etrm.

For example, if f is an id with etyp Hom(x,y) and g is an id with etyp Hom(y,z), then context(eg, :(g ∘ f)) computes the context [x,y,z,f,g].

The tricky thing comes from term formers like

weaken(x)::Term(n) ⊣ [n::Nat, x::Term(S(n))]

We get the ETyp for x from the e-graph, and then we have to ematch its argument with S(n) to figure out what n is... The problem is that in general S will not be injective, so this is ambiguous!

What we are going to do for now is say that types in the context of a term former can't be nested. I.e., we only allow types of the form Term(n), not Term(S(n)).

Fortunately, I don't think we care about any theories with this kind of context former.

Gatlab.Models.ModelInterface.Model — Type

A Julia value with type Model{T, Tuple{Ts...}} represents a model of the theory T that has the types Ts... assigned to the type constructors in T.

Note that we do not expect that structs subtyping Model will necessarily be singletons. In general, they may contain runtime data which is used in the implementations of the various methods. For instance, this would be the case for a slice category: the Model corresponding to a slice category would have runtime data of the object that we are slicing over.

An instance of Model{T, Tuple{Ts...}} should have overloads for

checkvalidity

ap

as described in the docstrings for each of these methods.

Gatlab.Models.ModelInterface.ap — Function

Meant to be overloaded as

ap(m::Model{T}, ::TrmTag{i}, args...; context_args...)

where:

i is the level of a term constructor in T

args consists of, for each argument to the term constructor i, a valid value

of the relevant type

context_args consists of keyword arguments for each of the elements of

the context that are not mentioned in args. Sometimes these are needed for implementation, sometimes they are not, so it is possible to call ap without instantiating all of these and hoping for the best.

Returns the result of applying the term constructor to the arguments, according to the model m. For instance, this is used to compose morphisms, or to get the identity morphism at an object.

Note that we assume in ap that the arguments have already been checked, using checkvalidity.

Gatlab.Models.ModelInterface.checkvalidity — Function

Meant to be overloaded as

checkvalidity(m::Model{T}, ::TypTag{i}, args..., x) where {T <: AbstractTheory} = ...

where:

i is the level of a type constructor in T

args are the pre-checked type arguments to the type constructor

x is the value to be checked

Returns a boolean which is true if x is a valid element of the type i with type arguments args... according to Model and false otherwise.

A basic implementation of checkvalidity might simply use dispatching on the type to return true. But in general, it might be necessary to verify runtime data, for instance checking that something is sorted or something is injective.

Gatlab.Models.ModelInterface.checkvalidity — Method

If not otherwise specified, we assume that a given value is not valid.

Gatlab.Dsl — Module

A module for various features of Gatlab packaged as domain specific languages