Standard Library

Gatlab.Stdlib.StdTheories.Categories.TypedHom — Type

Any implementor of TypedHom{Ob, Hom} should have precisely the fields

dom::Ob

codom::Ob

morphism::Hom

The reason this is not a struct is that we want to be able to control the name of the type.